Remove duplicate code. Fix references.
